+++ test/test_runner.h
@@ -42,4 +42,12 @@ void get_unit_c_tests(const struct CMUnitTest** out_tests, size_t* out_count);
  */
 void get_unit_c_null_tests(const struct CMUnitTest** out_tests, size_t* out_count);
 
+/**
+ * Get the processing-instruction test array from unit-c-pi.c.
+ *
+ * @param out_tests On success, set to pointer to the test array
+ * @param out_count On success, set to the number of tests
+ */
+void get_unit_c_pi_tests(const struct CMUnitTest** out_tests, size_t* out_count);
+
 #endif